ReInc 0.5.0: reimplemented gauntlet mode with few differences:
- shortened the length of each "section" from 6 floors to 8
- removed the mob spawn reduction
- added special rooms
- added guaranteed alchemy room every 4 depths
- reworked the economy: instead of fixed prices, the last mob drops 3-4x of average gold value for current floor
- reworded the descriptions
- damaging mobs in gauntlet mode now prolongs regeneration time, the time decreases with progression

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Thief.java
@@ -139,7 +139,7 @@ public class Thief extends Mob {
 
 	@Override
 	public int defenseProc(Char enemy, int damage) {
-		if (state == FLEEING) {
+		if (state == FLEEING && Dungeon.mode != Dungeon.GameMode.GAUNTLET) {
 			Dungeon.level.drop( new Gold(), pos ).sprite.drop();
 		}
